pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer website"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=21.2.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "validators>=0.22.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"
